Moved `test_access_esm1p5_manifest_exe_in_release_spack_location` test function from `src/model_config_tests` directory
